backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
openpyxl
duckdb
openai
